Divide structures
